backend/requirements.txt
cwltool
Django
djangorestframework
# gunicorn
Jinja2
kubernetes==28.1.0
loguru
psycopg2